// src/interfaces/Tenant.ts
export interface Tenant {
    id: number;
    description: string;
    email: string;
    addressLine1: string;
    addressLine2: string;
    addressCity: string;
    addressZipCode: string;
    document: string;
    phone: string;
    active: boolean;
    createdAt: string;
    updatedAt: string;
}
  